@@ -20,14 +20,17 @@
     "@jitsi/sdp-interop": "1.0.3",
     "@jitsi/sdp-simulcast": "0.3.0",
     "async": "0.9.0",
+    "base64-js": "1.3.1",
     "current-executing-script": "0.1.3",
     "jitsi-meet-logger": "github:jitsi/jitsi-meet-logger#5ec92357570dc8f0b7ffc1528820721c84c6af8b",
     "lodash.clonedeep": "4.5.0",
+    "lodash.debounce": "4.0.8",
     "lodash.isequal": "4.5.0",
     "sdp-transform": "2.3.0",
     "strophe.js": "1.3.4",
     "strophejs-plugin-disco": "0.0.2",
-    "strophejs-plugin-stream-management": "github:jitsi/strophejs-plugin-stream-management#e719a02b4f83856c1530882084a4b048ee622d45",
+    "strophejs-plugin-stream-management": "github:jitsi/strophejs-plugin-stream-management#001cf02bef2357234e1ac5d163611b4d60bf2b6a",
+    "uuid": "8.1.0",
     "webrtc-adapter": "7.5.0"
   },
   "devDependencies": {
@@ -40,15 +43,16 @@
     "babel-eslint": "10.0.1",
     "babel-loader": "8.0.4",
     "core-js": "2.5.1",
-    "eslint": "4.12.1",
-    "eslint-config-jitsi": "github:jitsi/eslint-config-jitsi#1.0.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
+    "eslint": "5.6.1",
+    "eslint-config-jitsi": "github:jitsi/eslint-config-jitsi#1.0.3",
+    "eslint-plugin-flowtype": "2.50.3",
+    "eslint-plugin-import": "2.20.2",
     "flow-bin": "0.104.0",
     "jasmine-core": "3.5.0",
-    "karma": "4.4.1",
+    "karma": "5.1.1",
     "karma-chrome-launcher": "3.1.0",
     "karma-jasmine": "3.1.1",
+    "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "4.0.2",
     "string-replace-loader": "2.1.1",
     "webpack": "4.43.0",
@@ -60,7 +64,8 @@
     "postinstall": "webpack -p",
     "test": "karma start karma.conf.js",
     "test-watch": "karma start karma.conf.js --no-single-run",
-    "validate": "npm ls"
+    "validate": "npm ls",
+    "watch": "webpack --config webpack.config.js --watch --mode development"
   },
   "main": "./index.js",
   "license": "Apache-2.0"
