@@ -1,15 +1,7 @@
 /* global __filename */
 
-import getActiveAudioDevice from './modules/detection/ActiveDeviceDetector';
-import AudioMixer from './modules/webaudio/AudioMixer';
-import * as DetectionEvents from './modules/detection/DetectionEvents';
-import TrackVADEmitter from './modules/detection/TrackVADEmitter';
-import { createGetUserMediaEvent } from './service/statistics/AnalyticsEvents';
-import AuthUtil from './modules/util/AuthUtil';
-import * as ConnectionQualityEvents
-    from './service/connectivity/ConnectionQualityEvents';
-import * as E2ePingEvents from './service/e2eping/E2ePingEvents';
-import GlobalOnErrorHandler from './modules/util/GlobalOnErrorHandler';
+import Logger from 'jitsi-meet-logger';
+
 import * as JitsiConferenceErrors from './JitsiConferenceErrors';
 import * as JitsiConferenceEvents from './JitsiConferenceEvents';
 import JitsiConnection from './JitsiConnection';
@@ -21,21 +13,31 @@ import JitsiTrackError from './JitsiTrackError';
 import * as JitsiTrackErrors from './JitsiTrackErrors';
 import * as JitsiTrackEvents from './JitsiTrackEvents';
 import * as JitsiTranscriptionStatus from './JitsiTranscriptionStatus';
-import LocalStatsCollector from './modules/statistics/LocalStatsCollector';
-import Logger from 'jitsi-meet-logger';
-import * as MediaType from './service/RTC/MediaType';
-import Resolutions from './service/RTC/Resolutions';
+import RTC from './modules/RTC/RTC';
+import browser from './modules/browser';
 import NetworkInfo from './modules/connectivity/NetworkInfo';
 import { ParticipantConnectionStatus }
     from './modules/connectivity/ParticipantConnectionStatus';
-import RTC from './modules/RTC/RTC';
-import browser from './modules/browser';
-import ScriptUtil from './modules/util/ScriptUtil';
-import recordingConstants from './modules/recording/recordingConstants';
+import getActiveAudioDevice from './modules/detection/ActiveDeviceDetector';
+import * as DetectionEvents from './modules/detection/DetectionEvents';
+import TrackVADEmitter from './modules/detection/TrackVADEmitter';
 import ProxyConnectionService
     from './modules/proxyconnection/ProxyConnectionService';
+import recordingConstants from './modules/recording/recordingConstants';
+import LocalStatsCollector from './modules/statistics/LocalStatsCollector';
+import precallTest from './modules/statistics/PrecallTest';
 import Statistics from './modules/statistics/statistics';
+import AuthUtil from './modules/util/AuthUtil';
+import GlobalOnErrorHandler from './modules/util/GlobalOnErrorHandler';
+import ScriptUtil from './modules/util/ScriptUtil';
 import * as VideoSIPGWConstants from './modules/videosipgw/VideoSIPGWConstants';
+import AudioMixer from './modules/webaudio/AudioMixer';
+import * as MediaType from './service/RTC/MediaType';
+import Resolutions from './service/RTC/Resolutions';
+import * as ConnectionQualityEvents
+    from './service/connectivity/ConnectionQualityEvents';
+import * as E2ePingEvents from './service/e2eping/E2ePingEvents';
+import { createGetUserMediaEvent } from './service/statistics/AnalyticsEvents';
 
 const logger = Logger.getLogger(__filename);
 
@@ -642,6 +644,8 @@ export default _mergeNamespaceAndModule({
             logger.debug('MediaStreamTrack contentHint attribute not supported');
         }
     },
+
+    precallTest,
 
     /* eslint-enable max-params */
 
