@@ -1,8 +1,8 @@
 import EventEmitter from 'events';
 
 import RTC from '../RTC/RTC';
-
 import { createAudioContext } from '../webaudio/WebAudioUtils';
+
 import { VAD_SCORE_PUBLISHED } from './DetectionEvents';
 
 /**
