@@ -1,10 +1,11 @@
+import { getLogger } from 'jitsi-meet-logger';
+
 import {
     TYPE_OPERATIONAL,
     TYPE_PAGE,
     TYPE_TRACK,
     TYPE_UI
 } from '../../service/statistics/AnalyticsEvents';
-import { getLogger } from 'jitsi-meet-logger';
 import browser from '../browser';
 
 const MAX_CACHE_SIZE = 100;
