@@ -1,7 +1,9 @@
 /* global __filename */
 import { getLogger } from 'jitsi-meet-logger';
-import * as JingleSessionState from './JingleSessionState';
+
 import Listenable from '../util/Listenable';
+
+import * as JingleSessionState from './JingleSessionState';
 
 const logger = getLogger(__filename);
 
