@@ -12,13 +12,18 @@ export default class LastRequestTracker {
     /**
      * Starts tracking requests on the given connection.
      *
+     * @param {XmppConnection} xmppConnection - The XMPP connection which manages the given {@code stropheConnection}.
      * @param {Object} stropheConnection - Strophe connection instance.
      */
-    startTracking(stropheConnection) {
+    startTracking(xmppConnection, stropheConnection) {
         const originalRawInput = stropheConnection.rawInput;
 
-        stropheConnection.rawInput = function(...args) {
-            this._lastSuccess = Date.now();
+        stropheConnection.rawInput = (...args) => {
+            // It's okay to use rawInput callback only once the connection has been established, otherwise it will
+            // treat 'item-not-found' or other connection error on websocket reconnect as successful stanza received.
+            if (xmppConnection.connected) {
+                this._lastSuccess = Date.now();
+            }
             originalRawInput.apply(stropheConnection, args);
         };
     }
