@@ -3,9 +3,10 @@
 import { getLogger } from 'jitsi-meet-logger';
 import { Strophe } from 'strophe.js';
 
+import XMPPEvents from '../../service/xmpp/XMPPEvents';
+
 import ChatRoom from './ChatRoom';
 import { ConnectionPluginListenable } from './ConnectionPlugin';
-import XMPPEvents from '../../service/xmpp/XMPPEvents';
 
 const logger = getLogger(__filename);
 
